pyproject.toml
[tool.poetry]
name = "alkebulan-backend"
version = "0.1.0"
description = ""
authors = ["Victor Momodu <victoramomodu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
anthropic = "^0.34.2"
boto3 = "^1.35.29"
fastapi = { extras = ["standard"], version = "^0.115.0" }
flashrank = "^0.2.9"
img2table = "^1.3.0"
langchain-anthropic = "^0.2.1"
langchain-community = "^0.3.1"
langchain-openai = "^0.2.1"
openai = "^1.50.2"
opensearch-py = "^2.7.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pdfminer = "^20191125"
pillow = "^10.4.0"
pyjwt = "^2.9.0"
python = ">=3.10,<3.13"
python-docx = "^1.1.2"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.12"
qdrant-client = "^1.11.3"
requests = "^2.32.3"
requests-aws4auth = "^1.3.1"
sqlalchemy = "^2.0.35"
sqlmodel = "^0.0.22"
unstructured = "^0.15.14"
unstructured-client = "^0.26.0"
websockets = "^13.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
